# server/docs/swagger.yaml
basePath: /
definitions:
...
  model.User:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        maxLength: 255
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      deleted_at:
        example: "null"
        type: string
      email:
        example: admin@example.com
        maxLength: 100
        type: string
      id:
        example: 1
        type: integer
      nickname:
        example: 管理员
        maxLength: 50
        type: string
      password:
        example: "123456"
        maxLength: 100
        minLength: 6
        type: string
      phone:
        example: "13800138000"
        maxLength: 20
        type: string
      status:
        description: 1:启用 2:禁用
        enum:
        - 1
        - 2
        example: 1
        type: integer
      tenant_id:
        description: 租户ID
        example: 1
        type: integer
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      username:
        example: admin
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 基于Gin + React的后台管理系统API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go-React-Admin API
  version: "1.0"
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口，验证用户名密码并返回JWT Token
      parameters:
      - description: 用户登录信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"登录成功","token":"string","userId":"uint"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"string"}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: '{"error":"用户名或密码错误"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"生成Token失败"}'
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - 用户管理
  /api/logout:
    post:
      consumes:
      - application/json
      description: 用户登出接口
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"登出成功"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 用户登出
      tags:
      - 用户管理
...
  /api/register:
    post:
      consumes:
      - application/json
      description: 用户注册接口，创建新用户
      parameters:
      - description: 用户注册信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"注册成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"string"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"注册失败"}'
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - 用户管理
...
  /api/user/info:
    get:
      consumes:
      - application/json
      description: 根据JWT Token获取当前登录用户的信息
      produces:
      - application/json
      responses:
        "200":
          description: '{"user":model.User}'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: '{"error":"无法获取用户信息"}'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: '{"error":"用户不存在"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户信息
      tags:
      - 用户管理
  /api/users:
    get:
      consumes:
      - application/json
      description: 获取所有用户的列表
      produces:
      - application/json
      responses:
        "200":
          description: '{"users":[]model.User}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"获取用户列表失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 用户创建信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"用户创建成功","user":model.User}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"请求参数错误"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"创建用户失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 创建用户
      tags:
      - 用户管理
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据用户ID删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"用户删除成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"无效的用户ID"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"删除用户失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 根据用户ID更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户更新信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"用户更新成功"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '{"error":"请求参数错误"}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: '{"error":"更新用户失败"}'
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户
      tags:
      - 用户管理
schemes:
- http
- https
swagger: "2.0"